+++ lib/cpp/src/thrift/protocol/TBinaryProtocol.h
@@ -174,18 +174,18 @@ public:
 
   void checkReadBytesAvailable(TSet& set) override
   {
-      trans_->checkReadBytesAvailable(set.size_ * getMinSerializedSize(set.elemType_));
+      trans_->checkReadBytesAvailable(static_cast<int64_t>(set.size_) * getMinSerializedSize(set.elemType_));
   }
 
   void checkReadBytesAvailable(TList& list) override
   {
-      trans_->checkReadBytesAvailable(list.size_ * getMinSerializedSize(list.elemType_));
+      trans_->checkReadBytesAvailable(static_cast<int64_t>(list.size_) * getMinSerializedSize(list.elemType_));
   }
 
   void checkReadBytesAvailable(TMap& map) override
   {
       int elmSize = getMinSerializedSize(map.keyType_) + getMinSerializedSize(map.valueType_);
-      trans_->checkReadBytesAvailable(map.size_ * elmSize);
+      trans_->checkReadBytesAvailable(static_cast<int64_t>(map.size_) * elmSize);
   }
 
 protected:
